Remove response integrity check of config.json and layer
This change is being made because the `Docker-Content-Digest` value
sent back by docker distribution is a hash digest from the image
manifest list when only the image manifest is accepted by the client.

This issue on docker distribution tracks the regression:
https://github.com/docker/distribution/issues/2395

Fixes #23

Issue #24 tracks reverting this change when the docker distribution
issue is resolved.