Remove the response integrity check
In commit 63129b1 I removed the check for the function that fetches
layers but did not remove it for the function that fetches a whole
image.

This change is being made because the `Docker-Content-Digest` value
sent back by docker distribution is a hash digest from the image
manifest list when only the image manifest is accepted by the client.

This issue on docker distribution tracks the regression:
https://github.com/docker/distribution/issues/2395

Fixes #23

Issue #24 tracks reverting this change when the docker distribution
issue is resolved.

This change also fixes a test failure caused by the tests data dir not
being included in the cabal source distribution.