Build against `hnix` `master`
The motivation for this branch is so that `hocker` and `dhall` can both be
used within the same dependency tree.

The context for this is that `dhall` requires `megaparsec-7.0.0` or newer
whereas the latest `hnix` on Hackage only supports up to `megaparsec < 7.0`.
Since `hocker` depends on `hnix` that means that `hocker` is incompatible with
`dhall` if `hocker` depends on an `hnix` from Hackage.

However, the `master` branch of `hnix` on GitHub supports `megaparsec-7.*`, so
this change updates `hocker` to build against that.

Note that this shouldn't be merged until `hnix` pushes out a new release to
Hackage.