Add support for manifest lists
Amazon ecr responds with manifest list irrespective of what we
mention in the Accept header. Added an `--arch` option, that
defaults to builtin arch, to fetch manifest & config files for
only the required arch.

Strips Authorization Header when following redirects, this is
necessary when making requests to AWS as well as curl's default
behaviour